requirements.txt
pygame
python-dotenv
click
tinkoff-investments
